Current Bugs
Icons:
- I would prefer to use the font-awesome icon set instead of the
foundation one. Especially because they have different types of each
icon (light & dark) and I think they are prettier designed.

Menu:
- Burger-icon on responsive devices should be $secondary-color
- logo should not be truncated
- bg color of links should be $secondary-color and active state primary

landingpage:
include the icons on top of the 4 features.

profile:
- the colored rating-points should also be visibly on small screens
- include the social icons
- remove the big spacing at the end of the comments
- there is a small margin to the right on mobile devices that makes in
horizontal-scrollable.

compare:
- ratings / References on opened items (Bruce lee is collapseable)
- include icons

backend:
the sidebar should be always opened (except for mobile screens). Maybe
the currently used off-canvas element was not the right chose for the
left-sidebar (I think for the right side it is ok for a tiny
„notification-are“ or something like this). Maybe compare it with the
left-sidebar.